Refactor auction metrics with macro
